pyproject.toml
[project]
name = "testsmith"
version = "0.1.0"
description = "Forge QA test cases from text and documents using LLMs"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "google-genai>=0.3.0",
    "typer>=0.12.0",
    "pypdf>=4.0.0",
    "python-docx>=1.1.0",
    "rich>=13.0.0",
]

[project.scripts]
testsmith = "testsmith.cli:app"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["testsmith*"]
